Rename AktivitetStatusForm til TS
